pyproject.toml
[project]
name = "yugayu"
version = "0.3.0"
description = "AI Lab Orchestrator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "cryptography>=46.0.5",
]

[project.scripts]
yugayu = "yugayu.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]